Update: sbt, scalajs, scala2.12, scala2.13, play. Add: twirl module, native and sjs cross builds. Remove: scala2.11
